Start refactoring commands tree
